+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 
+- Fixed metadata auto-fetching for unsupported URL domains in discovery
 - Fixed editor form resetting during bulk metadata refresh for discovery releases
 - Fixed particular strings not being translated on locale change